Extract functionallity in separate modules
