+++ lib/sanbase_web/graphql/schema/types/kafka_types.ex
@@ -1,0 +1,59 @@
+defmodule SanbaseWeb.Graphql.Schema.KafkaTypes do
+  use Absinthe.Schema.Notation
+
+  enum(:side_enum, values: [:buy, :sell])
+
+  object :exchange_market_depth do
+    field(:source, :string)
+    field(:symbol, :string)
+    field(:timestamp, :datetime)
+    field(:ask, :float)
+    field(:asks025_percent_depth, :float)
+    field(:asks025_percent_volume, :float)
+    field(:asks05_percent_depth, :float)
+    field(:asks05_percent_volume, :float)
+    field(:asks075_percent_depth, :float)
+    field(:asks075_percent_volume, :float)
+    field(:asks10_percent_depth, :float)
+    field(:asks10_percent_volume, :float)
+    field(:asks1_percent_depth, :float)
+    field(:asks1_percent_volume, :float)
+    field(:asks20_percent_depth, :float)
+    field(:asks20_percent_volume, :float)
+    field(:asks2_percent_depth, :float)
+    field(:asks2_percent_volume, :float)
+    field(:asks30_percent_depth, :float)
+    field(:asks30_percent_volume, :float)
+    field(:asks5_percent_depth, :float)
+    field(:asks5_percent_volume, :float)
+    field(:bid, :float)
+    field(:bids025_percent_depth, :float)
+    field(:bids025_percent_volume, :float)
+    field(:bids05_percent_depth, :float)
+    field(:bids05_percent_volume, :float)
+    field(:bids075_percent_depth, :float)
+    field(:bids075_percent_volume, :float)
+    field(:bids10_percent_depth, :float)
+    field(:bids10_percent_volume, :float)
+    field(:bids1_percent_depth, :float)
+    field(:bids1_percent_volume, :float)
+    field(:bids20_percent_depth, :float)
+    field(:bids20_percent_volume, :float)
+    field(:bids2_percent_depth, :float)
+    field(:bids2_percent_volume, :float)
+    field(:bids30_percent_depth, :float)
+    field(:bids30_percent_volume, :float)
+    field(:bids5_percent_depth, :float)
+    field(:bids5_percent_volume, :float)
+  end
+
+  object :exchange_trade do
+    field(:source, :string)
+    field(:symbol, :string)
+    field(:timestamp, :datetime)
+    field(:side, :side_enum)
+    field(:amount, :float)
+    field(:price, :float)
+    field(:cost, :float)
+  end
+end
